tests: Check return value of clutter_init_with_args instead of error
Some of the tests were ignoring the return value of
clutter_init_with_args and instead they would recognise an error by
seeing whether the GError parameter was set. This patch changes it to
check the return value so that it won't give a warning when
G_GNUC_WARN_UNUSED_RESULT is on that function in Clutter 1.8.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2574